+++ javaparser-core-testing-bdd/src/test/java/com/github/javaparser/CommentParsingTest.java
@@ -19,33 +19,24 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.bdd;
+package com.github.javaparser;
 
-import com.github.javaparser.bdd.steps.ParsingSteps;
-import com.github.javaparser.bdd.steps.SharedSteps;
+import com.github.javaparser.steps.CommentParsingSteps;
 import com.github.valfirst.jbehave.junit.monitoring.JUnitReportingRunner;
 import org.jbehave.core.steps.InjectableStepsFactory;
 import org.jbehave.core.steps.InstanceStepsFactory;
 import org.junit.runner.RunWith;
 
-import java.util.HashMap;
-import java.util.Map;
-
 @RunWith(JUnitReportingRunner.class)
-public class ParsingTest extends BasicJBehaveTest {
+public class CommentParsingTest extends BasicJBehaveTest {
 
     @Override
     public InjectableStepsFactory stepsFactory() {
-        Map<String, Object> state = new HashMap<>();
-
-        return new InstanceStepsFactory(configuration(),
-                new SharedSteps(state),
-                new ParsingSteps(state));
+        return new InstanceStepsFactory(configuration(), new CommentParsingSteps());
     }
 
-    public ParsingTest() {
-        super("**/bdd/parsing*.story");
+    public CommentParsingTest() {
+        super("**/comment*.story");
     }
 }
-
