+++ javaparser-core-testing-bdd/src/test/java/com/github/javaparser/ComparingTest.java
@@ -19,24 +19,33 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.bdd;
+package com.github.javaparser;
 
-import com.github.javaparser.bdd.steps.CommentParsingSteps;
+import com.github.javaparser.steps.ComparingSteps;
+import com.github.javaparser.steps.SharedSteps;
 import com.github.valfirst.jbehave.junit.monitoring.JUnitReportingRunner;
 import org.jbehave.core.steps.InjectableStepsFactory;
 import org.jbehave.core.steps.InstanceStepsFactory;
 import org.junit.runner.RunWith;
 
+import java.util.HashMap;
+import java.util.Map;
+
 @RunWith(JUnitReportingRunner.class)
-public class CommentParsingTest extends BasicJBehaveTest {
+public class ComparingTest extends BasicJBehaveTest {
 
     @Override
     public InjectableStepsFactory stepsFactory() {
-        return new InstanceStepsFactory(configuration(), new CommentParsingSteps());
+        Map<String, Object> state = new HashMap<>();
+
+        return new InstanceStepsFactory(configuration(),
+                new SharedSteps(state),
+                new ComparingSteps());
     }
 
-    public CommentParsingTest() {
-        super("**/bdd/comment*.story");
+    public ComparingTest() {
+        super("**/comparing*.story");
     }
 }
+
