+++ javaparser-core-testing-bdd/src/test/java/com/github/javaparser/TestUtils.java
@@ -19,14 +19,14 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.bdd;
+package com.github.javaparser;
 
 import java.io.InputStream;
 
 public class TestUtils {
 
     public static InputStream getSampleStream(String sampleName) {
-        InputStream is = TestUtils.class.getClassLoader().getResourceAsStream("com/github/javaparser/bdd/samples/"
+        InputStream is = TestUtils.class.getClassLoader().getResourceAsStream("com/github/javaparser/samples/"
                 + sampleName + ".java");
         if (is == null) {
             throw new RuntimeException("Example not found, check your test. Sample name: " + sampleName);
