+++ javaparser-core-testing-bdd/src/test/java/com/github/javaparser/VisitorTest.java
@@ -19,24 +19,33 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.bdd;
+package com.github.javaparser;
 
-import com.github.javaparser.bdd.steps.PrettyPrintingSteps;
+import com.github.javaparser.steps.SharedSteps;
+import com.github.javaparser.steps.VisitorSteps;
 import com.github.valfirst.jbehave.junit.monitoring.JUnitReportingRunner;
 import org.jbehave.core.steps.InjectableStepsFactory;
 import org.jbehave.core.steps.InstanceStepsFactory;
 import org.junit.runner.RunWith;
 
+import java.util.HashMap;
+import java.util.Map;
+
 @RunWith(JUnitReportingRunner.class)
-public class PrettyPrintingTest extends BasicJBehaveTest {
+public class VisitorTest extends BasicJBehaveTest {
 
     @Override
     public InjectableStepsFactory stepsFactory() {
-        return new InstanceStepsFactory(configuration(), new PrettyPrintingSteps());
+        Map<String, Object> state = new HashMap<>();
+        return new InstanceStepsFactory(configuration(),
+                new SharedSteps(state),
+                new VisitorSteps(state));
     }
 
-    public PrettyPrintingTest() {
-        super("**/bdd/pretty_printing*.story");
+    public VisitorTest() {
+        super("**/visitor*.story");
     }
 }
+
+
