+++ javaparser-core-testing-bdd/src/test/java/com/github/javaparser/steps/ComparingSteps.java
@@ -19,7 +19,7 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.bdd.steps;
+package com.github.javaparser.steps;
 
 import com.github.javaparser.ast.CompilationUnit;
 import org.jbehave.core.annotations.Given;
