+++ javaparser-core-testing-bdd/src/test/java/com/github/javaparser/steps/PositionRangeSteps.java
@@ -19,7 +19,7 @@
  * GNU Lesser General Public License for more details.
  */
 
-package com.github.javaparser.bdd.steps;
+package com.github.javaparser.steps;
 
 import com.github.javaparser.Position;
 import com.github.javaparser.Range;
