+++ javaparser-core-testing-bdd/src/test/resources/com/github/javaparser/comparing_scenarios.story
