+++ javaparser-core-testing-bdd/src/test/resources/com/github/javaparser/manipulation_scenarios.story
