+++ javaparser-core-testing-bdd/src/test/resources/com/github/javaparser/parsing_scenarios.story
