+++ javaparser-core-testing-bdd/src/test/resources/com/github/javaparser/position_range_scenarios.story
