+++ javaparser-core-testing-bdd/src/test/resources/com/github/javaparser/samples/ClassInLatin1.java
