+++ javaparser-core-testing-bdd/src/test/resources/com/github/javaparser/samples/JavaConcepts.java
