+++ pom.xml
@@ -3,6 +3,7 @@
     <modules>
         <module>javaparser-core</module>
         <module>javaparser-core-testing</module>
+        <module>javaparser-core-testing-bdd</module>
         <module>javaparser-core-generators</module>
         <module>javaparser-core-metamodel-generator</module>
         <module>javaparser-core-serialization</module>
